requirements.txt
streamlit
torch
torchvision
opencv-python
albumentations
scikit-learn
transformers
numpy
matplotlib
Pillow
tqdm
numpy
